requirements.txt
torch==2.3.1           # chọn CPU/CUDA theo môi trường
sentence-transformers==3.0.1
transformers==4.43.4
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
rapidfuzz==3.9.6

# I/O
pymupdf==1.24.9        # PDF
python-docx==1.1.2     # DOCX
pdf2image==1.17.0      # optional OCR
pytesseract==0.3.10    # optional OCR (cần Tesseract 5.x)
unidecode==1.3.8

# Service & config & report
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
jinja2==3.1.4